+++ backend/src/Enum/AuthenticationServerType.php
@@ -5,4 +5,5 @@ namespace App\Enum;
 enum AuthenticationServerType: string
 {
     case Ldap = 'ldap';
+    case Oidc = 'oidc';
 }